+++ resources/image.rb
@@ -10,6 +10,7 @@ property :force, [true, false], default: false, desired_state: false
 property :nocache, [true, false], default: false
 property :noprune, [true, false], default: false
 property :repo, String, name_property: true
+property :repo_credential_override, String
 property :rm, [true, false], default: true
 property :source, String
 property :tag, String, default: 'latest'
@@ -140,7 +141,7 @@ action_class do
 
   def pull_image
     with_retries do
-      creds = credentails
+      creds = credentials
       original_image = Docker::Image.get(image_identifier, {}, connection) if Docker::Image.exist?(image_identifier, {}, connection)
       new_image = Docker::Image.create({ 'fromImage' => image_identifier }, creds, connection)
 
@@ -150,7 +151,7 @@ action_class do
 
   def push_image
     with_retries do
-      creds = credentails
+      creds = credentials
       i = Docker::Image.get(image_identifier, {}, connection)
       i.push(creds, repo_tag: image_identifier)
     end
@@ -175,8 +176,25 @@ action_class do
     end
   end
 
-  def credentails
+  def credentials
+    return new_resource.repo_credential_override unless new_resource.repo_credential_override.nil? || new_resource.repo_credential_override.empty?
+
     registry_host = parse_registry_host(new_resource.repo)
-    node.run_state['docker_auth'] && node.run_state['docker_auth'][registry_host] || {}
+    # Return unless we have credentials stored and we have a custom repo path
+    return registry_host unless registry_host.include?('/') && node.run_state.key?('docker_auth')
+
+    # Longest-path registry credentials match
+    registry_host = nil
+    registry_split = new_resource.repo.split('/')
+    join_count = registry_split.count
+
+    until registry_host || (join_count < 0)
+      join_count -= 1
+      registry_host = registry_split[0..join_count].join('/') if node.run_state['docker_auth'].key?(registry_split[0..join_count].join('/'))
+    end
+
+    return {} unless registry_host
+
+    node.run_state.dig('docker_auth', registry_host) || {}
   end
 end